requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==4.2.20
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
